Use isHoneycomb/isOtel everywhere, including prelude
